added level select validation for previous level complete.
This code has the effect of making both age groups work at the same
time. To fix this i think we need to create a second results screen and
copy the code across and make the results variables more specific to
age group so we can then have an additional highscore page that selects
the age group and then once that is selected then it takes you to the
appropriate highscore page.